+++ ydb/core/health_check/builder.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include "builder_merge_issues.h"
+#include "builder_group_checker.h"
 #include "log.h"
 #include "health_check.h"
-#include "health_check_data.h"
-#include "health_check_helper.h"
+#include "health_check_structs.h"
+#include "health_check_utils.h"
 #include "self_check_request.h"
 
 #include <ydb/core/base/hive.h>
@@ -29,9 +31,9 @@
 #include <ydb/library/wilson_ids/wilson.h>
 #include <ydb/public/api/protos/ydb_monitoring.pb.h>
 
-#include <regex>
-
 namespace NKikimr::NHealthCheck {
+
+using namespace NSchemeShard;
 
 struct TOverallStateContext {
     Ydb::Monitoring::SelfCheckResult* Result;
@@ -87,16 +89,78 @@ template <typename Derived>
 class TBuilderBase {
 public:
     template <typename... TArgs>
-    decltype(auto) Build(TArgs&&... args) {
-        return static_cast<Derived*>(this)->BuildImpl(std::forward<TArgs>(args)...);
+    decltype(auto) Fill(TArgs&&... args) {
+        return static_cast<Derived*>(this)->FillImpl(std::forward<TArgs>(args)...);
     }
 };
 
-// each builder incapsulates neccesary data from /home/andrew-rykov/git/ydb/ydb/core/health_check/builder.cpp
-// and provides declarartion BuildImpl method to build the result, implementation in /home/andrew-rykov/git/ydb/ydb/core/health_check/builder.cpp
-template<typename TDatabase, typename TSystemTablets, typename TStorage>
-class TBuilderResult : TBuilderBase<TBuilderResult<TDatabase, TSystemTablets, TStorage>> {
+template<
+    typename TResult,
+    typename TDatabase,
+    typename TCompute,
+    typename TSystemTablets,
+    typename TTablets,
+    typename TStorage,
+    typename TStoragePool,
+    typename TStorageGroup,
+    typename TVDisk,
+    typename TPDisk,
+    typename TStorageGroupWB,
+    typename TVDiskWB,
+    typename TPDiskWB
+>
+class TBuilderContext {
 public:
+    THolder<TResult> Result;
+    THolder<TDatabase> Database;
+    THolder<TCompute> Compute;
+    THolder<TSystemTablets> SystemTablets;
+    THolder<TTablets> Tablets;
+    THolder<TStorage> Storage;
+    THolder<TStoragePool> StoragePool;
+    THolder<TStorageGroup> StorageGroup;
+    THolder<TVDisk> VDisk;
+    THolder<TPDisk> PDisk;
+    THolder<TStorageGroupWB> StorageGroupWithWhiteboard;
+    THolder<TVDiskWB> StorageVDiskWithWhiteboard;
+    THolder<TPDiskWB> StoragePDiskWithWhiteboard;
+
+    TBuilderContext(
+        THolder<TResult> result,
+        THolder<TDatabase> database,
+        THolder<TCompute> compute,
+        THolder<TSystemTablets> systemTablets,
+        THolder<TTablets> tablets,
+        THolder<TStorage> storage,
+        THolder<TStoragePool> storagePool,
+        THolder<TStorageGroup> storageGroup,
+        THolder<TVDisk> vdisk,
+        THolder<TPDisk> pdisk,
+        THolder<TStorageGroupWB> groupWB,
+        THolder<TVDiskWB> vdiskWB,
+        THolder<TPDiskWB> pdiskWB
+    )
+        : Result(std::move(result))
+        , Database(std::move(database))
+        , Compute(std::move(compute))
+        , SystemTablets(std::move(systemTablets))
+        , Tablets(std::move(tablets))
+        , Storage(std::move(storage))
+        , StoragePool(std::move(storagePool))
+        , StorageGroup(std::move(storageGroup))
+        , VDisk(std::move(vdisk))
+        , PDisk(std::move(pdisk))
+        , StorageGroupWithWhiteboard(std::move(groupWB))
+        , StorageVDiskWithWhiteboard(std::move(vdiskWB))
+        , StoragePDiskWithWhiteboard(std::move(pdiskWB))
+    {}
+};
+
+template<typename TContext>
+class TBuilderResult : TBuilderBase<TBuilderResult<TContext>>  {
+public:
+    THolder<TContext> Builders;
+
     TString DomainPath;
     TString FilterDatabase;
     bool HaveAllBSControllerInfo;
@@ -106,12 +170,7 @@ public:
     THashMap<TString, TStoragePoolState>& StoragePoolStateByName;
     THashSet<TString>& StoragePoolSeenByName;
 
-    THolder<TDatabase> BuilderDatabase;
-    THolder<TSystemTablets> BuilderSystemTablets;
-    THolder<TStorage> BuilderStorage;
-    TBuilderResult(THolder<TDatabase> builderDatabase,
-                   THolder<TSystemTablets> builderSystemTablets,
-                   THolder<TStorage> builderStorage,
+    TBuilderResult(THolder<TContext> builders,
                    const TString& domainPath,
                    const TString& filterDatabase,
                    const bool haveAllBSControllerInfo,
@@ -120,9 +179,7 @@ public:
                    THashSet<ui64>& storagePoolSeen,
                    THashMap<TString, TStoragePoolState>& storagePoolStateByName,
                    THashSet<TString>& storagePoolSeenByName)
-        : BuilderDatabase(builderDatabase)
-        , BuilderSystemTablets(builderSystemTablets)
-        , BuilderStorage(builderStorage)
+        : Builders(builders)
         , DomainPath(domainPath)
         , FilterDatabase(filterDatabase)
         , HaveAllBSControllerInfo(haveAllBSControllerInfo)
@@ -138,85 +195,170 @@ private:
         return FilterDatabase && FilterDatabase != DomainPath;
     }
 
-    void BuildImpl(TOverallStateContext context);
+    void FillImpl(TOverallStateContext context) {
+        if (IsSpecificDatabaseFilter()) {
+            Builders->Database.Fill(context, FilterDatabase, DatabaseState[FilterDatabase]);
+        } else {
+            for (auto& [path, state] : DatabaseState) {
+                Builders->Database.Fill(context, path, state);
+            }
+        }
+        if (DatabaseState.empty()) {
+            Ydb::Monitoring::DatabaseStatus& databaseStatus(*context.Result->add_database_status());
+            TSelfCheckResult tabletContext;
+            tabletContext.Location.mutable_database()->set_name(DomainPath);
+            databaseStatus.set_name(DomainPath);
+            {
+                TDatabaseState databaseState;
+                Builders->SystemTablets.Fill(databaseState, {&tabletContext, "SYSTEM_TABLET"});
+                context.UpdateMaxStatus(tabletContext.GetOverallStatus());
+            }
+        }
+        if (!FilterDatabase) {
+            TDatabaseState unknownDatabase;
+            bool fillUnknownDatabase = false;
+            if (HaveAllBSControllerInfo) {
+                for (auto& [id, pool] : StoragePoolState) {
+                    if (StoragePoolSeen.count(id) == 0) {
+                        unknownDatabase.StoragePools.insert(id);
+                    }
+                }
+                fillUnknownDatabase = !unknownDatabase.StoragePools.empty();
+            } else {
+                for (auto& [name, pool] : StoragePoolStateByName) {
+                    if (StoragePoolSeenByName.count(name) == 0) {
+                        unknownDatabase.StoragePoolNames.insert(name);
+                    }
+                }
+                fillUnknownDatabase = !unknownDatabase.StoragePoolNames.empty();
+            }
+            if (fillUnknownDatabase) {
+                Ydb::Monitoring::DatabaseStatus& databaseStatus(*context.Result->add_database_status());
+                TSelfCheckResult storageContext;
+                Builders->Storage.Fill(unknownDatabase, *databaseStatus.mutable_storage(), {&storageContext, "STORAGE"});
+                databaseStatus.set_overall(storageContext.GetOverallStatus());
+                context.UpdateMaxStatus(storageContext.GetOverallStatus());
+                context.AddIssues(storageContext.IssueRecords);
+            }
+        }
+        context.FillSelfCheckResult();
+    }
 };
 
-template<typename TCompute, typename TStorage>
-class TBuilderDatabase : TBuilderBase<TBuilderDatabase<TCompute, TStorage>> {
+template<typename TContext>
+class TBuilderDatabase : TBuilderBase<TBuilderDatabase<TContext>> {
 public:
-    THolder<TCompute> BuilderCompute;
-    THolder<TStorage> BuilderStorage;
+    THolder<TContext> Builders;
+
     bool IsSpecificDatabaseFilter;
 
-    TBuilderDatabase(THolder<TCompute> builderCompute, THolder<TStorage> builderStorage, const bool isSpecificDatabaseFilter)
-        : BuilderCompute(builderCompute)
-        , BuilderStorage(builderStorage)
+    TBuilderDatabase(THolder<TContext> builders, const bool isSpecificDatabaseFilter)
+        : Builders(builders)
         , IsSpecificDatabaseFilter(isSpecificDatabaseFilter)
     {}
 
-    void BuildImpl(TOverallStateContext& context, const TString& path, TDatabaseState& state);
+    void FillImpl(TOverallStateContext& context, const TString& path, TDatabaseState& state) {
+        Ydb::Monitoring::DatabaseStatus& databaseStatus(*context.Result->add_database_status());
+        TSelfCheckResult dbContext("DATABASE");
+        if (!IsSpecificDatabaseFilter) {
+            dbContext.Location.mutable_database()->set_name(path);
+        }
+        databaseStatus.set_name(path);
+        Builders.Compute.Fill(state, *databaseStatus.mutable_compute(), {&dbContext, "COMPUTE"});
+        Builders.Storage.Fill(state, *databaseStatus.mutable_storage(), {&dbContext, "STORAGE"});
+        if (databaseStatus.compute().overall() != Ydb::Monitoring::StatusFlag::GREEN
+                && databaseStatus.storage().overall() != Ydb::Monitoring::StatusFlag::GREEN) {
+            dbContext.ReportStatus(MaxStatus(databaseStatus.compute().overall(), databaseStatus.storage().overall()),
+                "Database has multiple issues", ETags::DBState, { ETags::ComputeState, ETags::StorageState});
+        } else if (databaseStatus.compute().overall() != Ydb::Monitoring::StatusFlag::GREEN) {
+            dbContext.ReportStatus(databaseStatus.compute().overall(), "Database has compute issues", ETags::DBState, {ETags::ComputeState});
+        } else if (databaseStatus.storage().overall() != Ydb::Monitoring::StatusFlag::GREEN) {
+            dbContext.ReportStatus(databaseStatus.storage().overall(), "Database has storage issues", ETags::DBState, {ETags::StorageState});
+        }
+        databaseStatus.set_overall(dbContext.GetOverallStatus());
+        context.UpdateMaxStatus(dbContext.GetOverallStatus());
+        context.AddIssues(dbContext.IssueRecords);
+        if (!context.HasDegraded && context.Status != Ydb::Monitoring::StatusFlag::GREEN && dbContext.HasTags({ETags::StorageState})) {
+            context.HasDegraded = true;
+        }
+    }
 };
 
-template<typename TComputeNode, typename TSystemTablets, typename TComputeDatabase, typename TTablets>
-class TBuilderCompute {
+template<typename TContext>
+class TBuilderCompute : TBuilderBase<TBuilderCompute<TContext>> {
 public:
+    THolder<TContext> Builders;
+
     THashMap<TSubDomainKey, TString> FilterDomainKey;
     THashMap<TNodeId, const NKikimrWhiteboard::TSystemStateInfo*> MergedNodeSystemState;
 
-    THolder<TComputeNode> BuilderComputeNode;
-    THolder<TSystemTablets> BuilderSystemTablets;
-    THolder<TComputeDatabase> BuilderComputeDatabase;
-    THolder<TTablets> BuilderTablets;
-
-    TBuilderCompute(THolder<TComputeNode> builderComputeNode,
-                    THolder<TSystemTablets> builderSystemTablets,
-                    THolder<TComputeDatabase> builderComputeDatabase,
-                    const TTablets& builderTablets,
+    TBuilderCompute(THolder<TContext> builders,
                     const THashMap<TSubDomainKey, TString>& filterDomainKey,
                     const THashMap<TNodeId, const NKikimrWhiteboard::TSystemStateInfo*> mergedNodeSystemState)
-        : BuilderComputeNode(builderComputeNode)
-        , BuilderSystemTablets(builderSystemTablets)
-        , BuilderComputeDatabase(builderComputeDatabase)
-        , BuilderTablets(builderTablets)
+        : Builders(builders)
         , FilterDomainKey(filterDomainKey)
         , MergedNodeSystemState(mergedNodeSystemState)
     {}
 
-    void BuildImpl(TDatabaseState& databaseState, Ydb::Monitoring::ComputeStatus& computeStatus, TSelfCheckResult context);
+    void FillImpl(TDatabaseState& databaseState, Ydb::Monitoring::ComputeStatus& computeStatus, TSelfCheckResult context);
 };
 
-template<typename TDescribe>
-class TBuilderComputeDatabase : public TBuilderBase<TBuilderComputeDatabase<TDescribe>> {
+class TBuilderComputeDatabase : public TBuilderBase<TBuilderComputeDatabase> {
 public:
-    TDescribe& DescribeByPath;
+    const THashMap<TString, TRequestResponse<TEvSchemeShard::TEvDescribeSchemeResult>>& DescribeByPath;
 
-    TBuilderComputeDatabase(const TDescribe& describeByPath)
+    TBuilderComputeDatabase(const THashMap<TString, TRequestResponse<TEvSchemeShard::TEvDescribeSchemeResult>>& describeByPath)
         : DescribeByPath(describeByPath)
     {}
 
-    void BuildImpl(TDatabaseState& databaseState, Ydb::Monitoring::ComputeStatus& computeStatus, TSelfCheckResult context);
+    void FillImpl(TDatabaseState& databaseState, Ydb::Monitoring::ComputeStatus& computeStatus, TSelfCheckResult context) {
+        auto itDescribe = DescribeByPath.find(databaseState.Path);
+        if (itDescribe != DescribeByPath.end() && itDescribe->second.IsOk()) {
+            const auto& domain(itDescribe->second->GetRecord().GetPathDescription().GetDomainDescription());
+            if (domain.GetPathsLimit() > 0) {
+                float usage = (float)domain.GetPathsInside() / domain.GetPathsLimit();
+                computeStatus.set_paths_quota_usage(usage);
+                if (static_cast<i64>(domain.GetPathsLimit()) - static_cast<i64>(domain.GetPathsInside()) <= 1) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Paths quota exhausted", ETags::QuotaUsage);
+                } else if (usage >= 0.99) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Paths quota usage is over 99%", ETags::QuotaUsage);
+                } else if (usage >= 0.90) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Paths quota usage is over 90%", ETags::QuotaUsage);
+                } else {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+                }
+            }
+            if (domain.GetShardsLimit() > 0) {
+                float usage = (float)domain.GetShardsInside() / domain.GetShardsLimit();
+                computeStatus.set_shards_quota_usage(usage);
+                if (static_cast<i64>(domain.GetShardsLimit()) - static_cast<i64>(domain.GetShardsInside()) <= 1) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Shards quota exhausted", ETags::QuotaUsage);
+                } else if (usage >= 0.99) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Shards quota usage is over 99%", ETags::QuotaUsage);
+                } else if (usage >= 0.90) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Shards quota usage is over 90%", ETags::QuotaUsage);
+                } else {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+                }
+            }
+        }
+    }
 };
 
-template<typename TSystemTablets, typename TComputeDatabase, typename TTablets>
-class TBuilderComputeNode : public TBuilderBase<TBuilderComputeNode<TSystemTablets, TComputeDatabase, TTablets>> {
-    THolder<TSystemTablets> BuilderSystemTablets;
-    THolder<TComputeDatabase> BuilderComputeDatabase;
-    THolder<TTablets> BuilderTablets;
+template<typename TContext>
+class TBuilderComputeNode : public TBuilderBase<TBuilderComputeNode<TContext>> {
+    THolder<TContext> Builders;
 public:
-    TBuilderComputeNode(THolder<TSystemTablets> builderSystemTablets,
-                        THolder<TComputeDatabase> builderComputeDatabase,
-                        THolder<TTablets> builderTablets)
-        : BuilderSystemTablets(builderSystemTablets)
-        , BuilderComputeDatabase(builderComputeDatabase)
-        , BuilderTablets(builderTablets)
+    TBuilderComputeNode(THolder<TContext> builders)
+        : Builders(builders)
     {}
 
-    void BuildImpl(TDatabaseState& databaseState, TNodeId nodeId, Ydb::Monitoring::ComputeNodeStatus& computeNodeStatus, TSelfCheckResult context);
+    void FillImpl(TDatabaseState& databaseState, TNodeId nodeId, Ydb::Monitoring::ComputeNodeStatus& computeNodeStatus, TSelfCheckResult context);
 };
 
 class TBuilderTablets : public TBuilderBase<TBuilderTablets> {
 public:
-    void BuildImpl(TDatabaseState& databaseState, TNodeId nodeId, Ydb::Monitoring::ComputeTabletStatus& computeTabletStatus, TSelfCheckResult context);
+    void FillImpl(TDatabaseState& databaseState, TNodeId nodeId, Ydb::Monitoring::ComputeTabletStatus& computeTabletStatus, TSelfCheckResult context);
 };
 
 class TBuilderSystemTablets : public TBuilderBase<TBuilderSystemTablets> {
@@ -226,89 +368,611 @@ public:
         : TabletRequests(tabletRequests)
     {}
 
-    Ydb::Monitoring::StatusFlag::Status BuildImpl(TDatabaseState& databaseState, TSelfCheckResult context);
+    Ydb::Monitoring::StatusFlag::Status FillImpl(TDatabaseState& databaseState, TSelfCheckResult context);
 };
 
-template<typename TPool>
-class TBuilderStorage : public TBuilderBase<TBuilderStorage<TPool>> {
+template<typename TContext>
+class TBuilderStorage : public TBuilderBase<TBuilderStorage<TContext>> {
 public:
-    THolder<TPool> BuilderPool;
+    THolder<TContext> Builders;
 
-    TBuilderStorage(THolder<TPool> builderPool)
-        : BuilderPool(builderPool)
+    bool HaveAllBSControllerInfo;
+    THashMap<ui64, TStoragePoolState>& StoragePoolState;
+    THashSet<ui64>& StoragePoolSeen;
+    THashMap<TString, TStoragePoolState>& StoragePoolStateByName;
+    THashSet<TString>& StoragePoolSeenByName;
+
+    TBuilderStorage(THolder<TContext> builders,
+                    bool haveAllBSControllerInfo,
+                    THashMap<ui64, TStoragePoolState>& storagePoolState,
+                    THashSet<ui64>& storagePoolSeen,
+                    THashMap<TString, TStoragePoolState>& storagePoolStateByName,
+                    THashSet<TString>& storagePoolSeenByName)
+        : Builders(builders)
+        , HaveAllBSControllerInfo(haveAllBSControllerInfo)
+        , StoragePoolState(storagePoolState)
+        , StoragePoolSeen(storagePoolSeen)
+        , StoragePoolStateByName(storagePoolStateByName)
+        , StoragePoolSeenByName(storagePoolSeenByName)
     {}
 
-    void BuildImpl(TDatabaseState& databaseState, Ydb::Monitoring::StorageStatus& storageStatus, TSelfCheckResult context);
+    void FillImpl(TDatabaseState& databaseState, Ydb::Monitoring::StorageStatus& storageStatus, TSelfCheckResult context) {
+        if (HaveAllBSControllerInfo && databaseState.StoragePools.empty()) {
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "There are no storage pools", ETags::StorageState);
+        } else {
+            if (HaveAllBSControllerInfo) {
+                for (const ui64 poolId : databaseState.StoragePools) {
+                    auto itStoragePoolState = StoragePoolState.find(poolId);
+                    if (itStoragePoolState != StoragePoolState.end() && itStoragePoolState->second.Groups) {
+                        Builders.Pool.Fill(itStoragePoolState->second, *storageStatus.add_pools(), {&context, "STORAGE_POOL"});
+                        StoragePoolSeen.emplace(poolId);
+                    }
+                }
+            } else {
+                for (const TString& poolName : databaseState.StoragePoolNames) {
+                    auto itStoragePoolState = StoragePoolStateByName.find(poolName);
+                    if (itStoragePoolState != StoragePoolStateByName.end()) {
+                        Builders.Pool.Fill(itStoragePoolState->second, *storageStatus.add_pools(), {&context, "STORAGE_POOL"});
+                        StoragePoolSeenByName.emplace(poolName);
+                    }
+                }
+            }
+            switch (context.GetOverallStatus()) {
+                case Ydb::Monitoring::StatusFlag::BLUE:
+                case Ydb::Monitoring::StatusFlag::YELLOW:
+                    context.ReportStatus(context.GetOverallStatus(), "Storage degraded", ETags::StorageState, {ETags::PoolState});
+                    break;
+                case Ydb::Monitoring::StatusFlag::ORANGE:
+                    context.ReportStatus(context.GetOverallStatus(), "Storage has no redundancy", ETags::StorageState, {ETags::PoolState});
+                    break;
+                case Ydb::Monitoring::StatusFlag::RED:
+                    context.ReportStatus(context.GetOverallStatus(), "Storage failed", ETags::StorageState, {ETags::PoolState});
+                    break;
+                default:
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+                    break;
+            }
+            if (!HaveAllBSControllerInfo) {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "System tablet BSC didn't provide information", ETags::StorageState);
+            }
+        }
+        if (databaseState.StorageQuota > 0) {
+            auto usage = (float)databaseState.StorageUsage / databaseState.StorageQuota;
+            if (usage > 0.9) {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Storage usage over 90%", ETags::StorageState);
+            } else if (usage > 0.85) {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Storage usage over 85%", ETags::StorageState);
+            } else if (usage > 0.75) {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Storage usage over 75%", ETags::StorageState);
+            }
+        }
+        storageStatus.set_overall(context.GetOverallStatus());
+    }
 };
 
-template<typename TGroup>
-class TBuilderStoragePool : public TBuilderBase<TBuilderStoragePool<TGroup>> {
+template<typename TContext>
+class TBuilderStoragePool : public TBuilderBase<TBuilderStoragePool<TContext>> {
 public:
-    THolder<TGroup> BuilderGroup;
+    THolder<TContext> Builders;
 
-    TBuilderStoragePool(THolder<TGroup> builderGroup)
-        : BuilderGroup(builderGroup)
+    bool HaveAllBSControllerInfo;
+    THashSet<TNodeId>& UnknownStaticGroups;
+    std::unordered_map<TGroupId, const NKikimrWhiteboard::TBSGroupStateInfo*>& MergedBSGroupState;
+    bool NeedMergeRecords; // Request->Request.merge_records()
+
+    TBuilderStoragePool(THolder<TContext> builders, bool haveAllBSControllerInfo,
+                        THashSet<TNodeId>& unknownStaticGroups,
+                        std::unordered_map<TGroupId, const NKikimrWhiteboard::TBSGroupStateInfo*>& mergedBSGroupState,
+                        bool needMergeRecords)
+        : Builders(builders)
+        , HaveAllBSControllerInfo(haveAllBSControllerInfo)
+        , UnknownStaticGroups(unknownStaticGroups)
+        , MergedBSGroupState(mergedBSGroupState)
+        , NeedMergeRecords(needMergeRecords)
     {}
 
-    void BuildImpl(const TStoragePoolState& pool, Ydb::Monitoring::StoragePoolStatus& storagePoolStatus, TSelfCheckResult context);
+    void MergeRecords(TList<TSelfCheckResult::TIssueRecord>& records) {
+        TMergeIssuesContext mergeContext(records);
+        if (NeedMergeRecords) {
+            mergeContext.MergeLevelRecords(ETags::GroupState);
+            mergeContext.MergeLevelRecords(ETags::VDiskState, ETags::GroupState);
+            mergeContext.MergeLevelRecords(ETags::PDiskState, ETags::VDiskState);
+        }
+        mergeContext.FillRecords(records);
+    }
+
+    void FillImpl(const TStoragePoolState& pool, Ydb::Monitoring::StoragePoolStatus& storagePoolStatus, TSelfCheckResult context) {
+        context.Location.mutable_storage()->mutable_pool()->set_name(pool.Name);
+        storagePoolStatus.set_id(pool.Name);
+        for (auto groupId : pool.Groups) {
+            if (HaveAllBSControllerInfo && !UnknownStaticGroups.contains(groupId)) {
+                Builders.StorageGroup.Fill(groupId, *storagePoolStatus.add_groups(), {&context, "STORAGE_GROUP"});
+            } else if (IsStaticGroup(groupId)) {
+                auto itGroup = MergedBSGroupState.find(groupId);
+                if (itGroup != MergedBSGroupState.end()) {
+                    Builders.StorageGroupWithWhiteboard.Fill(groupId, *itGroup->second, *storagePoolStatus.add_groups(), {&context, "STORAGE_GROUP"});
+                }
+            }
+        }
+
+        MergeRecords(context.IssueRecords);
+
+        switch (context.GetOverallStatus()) {
+            case Ydb::Monitoring::StatusFlag::BLUE:
+            case Ydb::Monitoring::StatusFlag::YELLOW:
+                context.ReportStatus(context.GetOverallStatus(), "Pool degraded", ETags::PoolState, {ETags::GroupState});
+                break;
+            case Ydb::Monitoring::StatusFlag::ORANGE:
+                context.ReportStatus(context.GetOverallStatus(), "Pool has no redundancy", ETags::PoolState, {ETags::GroupState});
+                break;
+            case Ydb::Monitoring::StatusFlag::RED:
+                context.ReportStatus(context.GetOverallStatus(), "Pool failed", ETags::PoolState, {ETags::GroupState});
+                break;
+            default:
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+                break;
+        }
+        storagePoolStatus.set_overall(context.GetOverallStatus());
+    }
 };
 
-template<typename TVDisk>
-class TBuilderStorageGroup : public TBuilderBase<TBuilderStorageGroup<TVDisk>> {
+template<typename TContext>
+class TBuilderStorageGroup : public TBuilderBase<TBuilderStorageGroup<TContext>> {
 public:
-    THolder<TVDisk> BuilderVDisk;
+    THolder<TContext> Builders;
 
-    TBuilderStorageGroup(THolder<TVDisk> builderVDisk)
-        : BuilderVDisk(builderVDisk)
+    THashMap<ui32, TGroupState>& GroupState;
+    std::unordered_map<TString, Ydb::Monitoring::StatusFlag::Status>& VDiskStatuses;
+
+    TBuilderStorageGroup(THolder<TContext> builders, THashMap<ui32, TGroupState>& groupState,
+                         std::unordered_map<TString, Ydb::Monitoring::StatusFlag::Status>& vDiskStatuses)
+        : Builders(builders)
+        , GroupState(groupState)
+        , VDiskStatuses(vDiskStatuses)
     {}
 
-    void BuildImpl(TGroupId groupId, Ydb::Monitoring::StorageGroupStatus& storageGroupStatus, TSelfCheckResult context);
+    void FillImpl(TGroupId groupId, Ydb::Monitoring::StorageGroupStatus& storageGroupStatus, TSelfCheckResult context) {
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_id()->Clear();
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->add_id(ToString(groupId));
+        storageGroupStatus.set_id(ToString(groupId));
+
+        auto itGroup = GroupState.find(groupId);
+        if (itGroup == GroupState.end()) {
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "Group has no vslots", ETags::GroupState);
+            storageGroupStatus.set_overall(context.GetOverallStatus());
+            return;
+        }
+
+        TGroupChecker checker(itGroup->second.ErasureSpecies, itGroup->second.LayoutCorrect);
+        const auto& slots = itGroup->second.VSlots;
+        for (const auto* slot : slots) {
+            const auto& slotInfo = slot->GetInfo();
+            auto slotId = GetVSlotId(slot->GetKey());
+            auto [statusIt, inserted] = VDiskStatuses.emplace(slotId, Ydb::Monitoring::StatusFlag::UNSPECIFIED);
+            if (inserted) {
+                Ydb::Monitoring::StorageVDiskStatus& vDiskStatus = *storageGroupStatus.add_vdisks();
+                Builders.VDisk.Fill(slot, vDiskStatus, {&context, "VDISK"});
+                statusIt->second = vDiskStatus.overall();
+            }
+            checker.AddVDiskStatus(statusIt->second, slotInfo.GetFailRealm());
+        }
+
+        context.Location.mutable_storage()->clear_node(); // group doesn't have node
+        context.OverallStatus = MinStatus(context.OverallStatus, Ydb::Monitoring::StatusFlag::YELLOW);
+        checker.ReportStatus(context);
+
+        storageGroupStatus.set_overall(context.GetOverallStatus());
+    }
 };
 
-template<typename TVDisk>
-class TBuilderStorageGroupWithWhiteboard : public TBuilderBase<TBuilderStorageGroupWithWhiteboard<TVDisk>> {
+template<typename TContext>
+class TBuilderStorageGroupWithWhiteboard : public TBuilderBase<TBuilderStorageGroupWithWhiteboard<TContext>> {
 public:
-    THolder<TVDisk> BuilderVDisk;
+    THolder<TContext> Builders;
 
-    TBuilderStorageGroupWithWhiteboard(THolder<TVDisk> builderVDisk)
-        : BuilderVDisk(builderVDisk)
+    std::unordered_map<TString, const NKikimrWhiteboard::TVDiskStateInfo*>& MergedVDiskState;
+    const THashMap<TNodeId, const TEvInterconnect::TNodeInfo*>& MergedNodeInfo;
+
+    TBuilderStorageGroupWithWhiteboard(THolder<TContext> builders,
+                                       std::unordered_map<TString, const NKikimrWhiteboard::TVDiskStateInfo*>& mergedVDiskState,
+                                       const THashMap<TNodeId, const TEvInterconnect::TNodeInfo*>& mergedNodeInfo)
+        : Builders(builders)
+        , MergedVDiskState(mergedVDiskState)
+        , MergedNodeInfo(mergedNodeInfo)
     {}
 
-    void BuildImpl(TGroupId groupId, const NKikimrWhiteboard::TBSGroupStateInfo& groupInfo, Ydb::Monitoring::StorageGroupStatus& storageGroupStatus, TSelfCheckResult context);
+    void FillImpl(TGroupId groupId, const NKikimrWhiteboard::TBSGroupStateInfo& groupInfo, Ydb::Monitoring::StorageGroupStatus& storageGroupStatus, TSelfCheckResult context) {
+        if (context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_id()->empty()) {
+            context.Location.mutable_storage()->mutable_pool()->mutable_group()->add_id();
+        }
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->set_id(0, ToString(groupId));
+        storageGroupStatus.set_id(ToString(groupId));
+        TGroupChecker checker(groupInfo.erasurespecies());
+        for (const auto& protoVDiskId : groupInfo.vdiskids()) {
+            TString vDiskId = GetVDiskId(protoVDiskId);
+            auto itVDisk = MergedVDiskState.find(vDiskId);
+            const TEvInterconnect::TNodeInfo* nodeInfo = nullptr;
+            if (itVDisk != MergedVDiskState.end()) {
+                TNodeId nodeId = itVDisk->second->nodeid();
+                auto itNodeInfo = MergedNodeInfo.find(nodeId);
+                if (itNodeInfo != MergedNodeInfo.end()) {
+                    nodeInfo = itNodeInfo->second;
+                }
+                context.Location.mutable_storage()->mutable_node()->set_id(nodeId);
+            } else {
+                context.Location.mutable_storage()->mutable_node()->clear_id();
+            }
+            if (nodeInfo) {
+                context.Location.mutable_storage()->mutable_node()->set_host(nodeInfo->Host);
+                context.Location.mutable_storage()->mutable_node()->set_port(nodeInfo->Port);
+            } else {
+                context.Location.mutable_storage()->mutable_node()->clear_host();
+                context.Location.mutable_storage()->mutable_node()->clear_port();
+            }
+            Ydb::Monitoring::StorageVDiskStatus& vDiskStatus = *storageGroupStatus.add_vdisks();
+            Builders.StorageVDiskWithWhiteboard.Fill(vDiskId, itVDisk != MergedVDiskState.end() ? *itVDisk->second : NKikimrWhiteboard::TVDiskStateInfo(), vDiskStatus, {&context, "VDISK"});
+            checker.AddVDiskStatus(vDiskStatus.overall(), protoVDiskId.ring());
+        }
+
+        context.Location.mutable_storage()->clear_node(); // group doesn't have node
+        context.OverallStatus = MinStatus(context.OverallStatus, Ydb::Monitoring::StatusFlag::YELLOW);
+        checker.ReportStatus(context);
+
+        BLOG_D("Group " << groupId << " has status " << context.GetOverallStatus());
+        storageGroupStatus.set_overall(context.GetOverallStatus());
+    }
 };
 
-template<typename TPDisk>
-class TBuilderStorageVDisk : public TBuilderBase<TBuilderStorageVDisk<TPDisk>> {
+template<typename TContext>
+class TBuilderStorageVDisk : public TBuilderBase<TBuilderStorageVDisk<TContext>> {
 public:
-    THolder<TPDisk> BuilderPDisk;
+    THolder<TContext> Builders;
 
-    TBuilderStorageVDisk(THolder<TPDisk> builderPDisk)
-        : BuilderPDisk(builderPDisk)
+    THashMap<TNodeId, const TEvInterconnect::TNodeInfo*>& MergedNodeInfo;
+
+    TBuilderStorageVDisk(THolder<TContext> builders,
+                         const THashMap<TNodeId, const TEvInterconnect::TNodeInfo*>& mergedNodeInfo)
+        : Builders(builders)
+        , MergedNodeInfo(mergedNodeInfo)
     {}
 
-    void BuildImpl(const NKikimrSysView::TVSlotEntry* vSlot, Ydb::Monitoring::StorageVDiskStatus& storageVDiskStatus, TSelfCheckResult context);
+    void FillImpl(const NKikimrSysView::TVSlotEntry* vSlot, Ydb::Monitoring::StorageVDiskStatus& storageVDiskStatus, TSelfCheckResult context) {
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_id()->Clear();
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->clear_id(); // you can see VDisks Group Id in vSlotId field
+
+        TNodeId nodeId = vSlot->GetKey().GetNodeId();
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->add_id(GetVDiskId(vSlot->GetInfo()));
+        context.Location.mutable_storage()->mutable_node()->set_id(nodeId);
+
+        auto itNodeInfo = MergedNodeInfo.find(nodeId);
+        if (itNodeInfo != MergedNodeInfo.end()) {
+            context.Location.mutable_storage()->mutable_node()->set_host(itNodeInfo->second->Host);
+            context.Location.mutable_storage()->mutable_node()->set_port(itNodeInfo->second->Port);
+        } else {
+            context.Location.mutable_storage()->mutable_node()->clear_host();
+            context.Location.mutable_storage()->mutable_node()->clear_port();
+        }
+
+        storageVDiskStatus.set_id(GetVSlotId(vSlot->GetKey()));
+
+        const auto& vSlotInfo = vSlot->GetInfo();
+
+        if (!vSlotInfo.HasStatusV2()) {
+            // this should mean that BSC recently restarted and does not have accurate data yet - we should not report to avoid false positives
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+            storageVDiskStatus.set_overall(context.GetOverallStatus());
+            return;
+        }
+
+        const auto *descriptor = NKikimrBlobStorage::EVDiskStatus_descriptor();
+        auto status = descriptor->FindValueByName(vSlot->GetInfo().GetStatusV2());
+        if (!status) { // this case is not expected because becouse bsc assignes status according EVDiskStatus enum
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "System tablet BSC didn't provide known status", ETags::VDiskState);
+            storageVDiskStatus.set_overall(context.GetOverallStatus());
+            return;
+        }
+
+        if (vSlot->GetKey().HasPDiskId()) {
+            TString pDiskId = GetPDiskId(vSlot->GetKey());
+            FillPDiskStatus(pDiskId, *storageVDiskStatus.mutable_pdisk(), {&context, "PDISK"});
+        }
+
+        if (status->number() == NKikimrBlobStorage::ERROR) {
+            // the disk is not operational at all
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "VDisk is not available", ETags::VDiskState,{ETags::PDiskState});
+            storageVDiskStatus.set_overall(context.GetOverallStatus());
+            return;
+        }
+
+        if (vSlotInfo.HasIsThrottling() && vSlotInfo.GetIsThrottling()) {
+            // throttling is active
+            auto message = TStringBuilder() << "VDisk is being throttled, rate "
+                << vSlotInfo.GetThrottlingRate() << " per mille";
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, message, ETags::VDiskState);
+            storageVDiskStatus.set_overall(context.GetOverallStatus());
+            return;
+        }
+
+        switch (status->number()) {
+            case NKikimrBlobStorage::REPLICATING: { // the disk accepts queries, but not all the data was replicated
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::BLUE, TStringBuilder() << "Replication in progress", ETags::VDiskState);
+                storageVDiskStatus.set_overall(context.GetOverallStatus());
+                return;
+            }
+            case NKikimrBlobStorage::INIT_PENDING:
+            case NKikimrBlobStorage::READY: { // the disk is fully operational and does not affect group fault tolerance
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+            }
+            default:
+                break;
+        }
+
+        context.ReportWithMaxChildStatus("VDisk have space issue",
+                            ETags::VDiskState,
+                            {ETags::PDiskSpace});
+
+        storageVDiskStatus.set_overall(context.GetOverallStatus());
+    }
 };
 
-template<typename TPDisk>
-class TBuilderStorageVDiskWithWhiteboard : public TBuilderBase<TBuilderStorageVDiskWithWhiteboard<TPDisk>> {
+template<typename TContext>
+class TBuilderStorageVDiskWithWhiteboard : public TBuilderBase<TBuilderStorageVDiskWithWhiteboard<TContext>> {
 public:
-    THolder<TPDisk> BuilderPDisk;
+    THolder<TContext> Builders;
 
-    TBuilderStorageVDiskWithWhiteboard(THolder<TPDisk> builderPDisk)
-        : BuilderPDisk(builderPDisk)
+    std::unordered_map<TString, const NKikimrWhiteboard::TPDiskStateInfo*>& MergedPDiskState;
+
+    TBuilderStorageVDiskWithWhiteboard(THolder<TContext> builders,
+                                       std::unordered_map<TString, const NKikimrWhiteboard::TPDiskStateInfo*>& mergedPDiskState)
+        : Builders(builders)
+        , MergedPDiskState(mergedPDiskState)
     {}
 
-    void BuildImpl(const TString& vDiskId, const NKikimrWhiteboard::TVDiskStateInfo& vDiskInfo, Ydb::Monitoring::StorageVDiskStatus& storageVDiskStatus, TSelfCheckResult context);
+    void FillImpl(const TString& vDiskId, const NKikimrWhiteboard::TVDiskStateInfo& vDiskInfo, Ydb::Monitoring::StorageVDiskStatus& storageVDiskStatus, TSelfCheckResult context) {
+        if (context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_id()->empty()) {
+            context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->add_id();
+        }
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->set_id(0, vDiskId);
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->clear_id(); // you can see VDisks Group Id in vDiskId field
+        storageVDiskStatus.set_id(vDiskId);
+        TString pDiskId = GetPDiskId(vDiskInfo);
+        auto itPDisk = MergedPDiskState.find(pDiskId);
+        if (itPDisk != MergedPDiskState.end()) {
+            Builders.StoragePDiskWithWhiteboard.Fill(pDiskId, *itPDisk->second, *storageVDiskStatus.mutable_pdisk(), {&context, "PDISK"});
+        }
+
+        if (!vDiskInfo.HasVDiskState()) {
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
+                                    TStringBuilder() << "VDisk is not available",
+                                    ETags::VDiskState,
+                                    {ETags::PDiskState});
+            storageVDiskStatus.set_overall(context.GetOverallStatus());
+            return;
+        }
+
+        switch (vDiskInfo.GetVDiskState()) {
+            case NKikimrWhiteboard::EVDiskState::OK:
+            case NKikimrWhiteboard::EVDiskState::Initial:
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+                break;
+            case NKikimrWhiteboard::EVDiskState::SyncGuidRecovery:
+                context.IssueRecords.clear();
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW,
+                                        TStringBuilder() << "VDisk state is " << NKikimrWhiteboard::EVDiskState_Name(vDiskInfo.GetVDiskState()),
+                                        ETags::VDiskState);
+                storageVDiskStatus.set_overall(context.GetOverallStatus());
+                return;
+            case NKikimrWhiteboard::EVDiskState::LocalRecoveryError:
+            case NKikimrWhiteboard::EVDiskState::SyncGuidRecoveryError:
+            case NKikimrWhiteboard::EVDiskState::PDiskError:
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
+                                        TStringBuilder() << "VDisk state is " << NKikimrWhiteboard::EVDiskState_Name(vDiskInfo.GetVDiskState()),
+                                        ETags::VDiskState,
+                                        {ETags::PDiskState});
+                storageVDiskStatus.set_overall(context.GetOverallStatus());
+                return;
+        }
+
+        if (!vDiskInfo.GetReplicated()) {
+            context.IssueRecords.clear();
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::BLUE, "Replication in progress", ETags::VDiskState);
+            storageVDiskStatus.set_overall(context.GetOverallStatus());
+            return;
+        }
+
+        if (vDiskInfo.HasDiskSpace()) {
+            switch(vDiskInfo.GetDiskSpace()) {
+                case NKikimrWhiteboard::EFlag::Green:
+                    if (context.IssueRecords.size() == 0) {
+                        context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+                    } else {
+                        context.ReportStatus(context.IssueRecords.begin()->IssueLog.status(),
+                                            TStringBuilder() << "VDisk is degraded",
+                                            ETags::VDiskState,
+                                            {ETags::PDiskSpace});
+                    }
+                    break;
+                case NKikimrWhiteboard::EFlag::Red:
+                    context.ReportStatus(GetFlagFromWhiteboardFlag(vDiskInfo.GetDiskSpace()),
+                                            TStringBuilder() << "DiskSpace is " << NKikimrWhiteboard::EFlag_Name(vDiskInfo.GetDiskSpace()),
+                                            ETags::VDiskState,
+                                            {ETags::PDiskSpace});
+                    break;
+                default:
+                    context.ReportStatus(GetFlagFromWhiteboardFlag(vDiskInfo.GetDiskSpace()),
+                                            TStringBuilder() << "DiskSpace is " << NKikimrWhiteboard::EFlag_Name(vDiskInfo.GetDiskSpace()),
+                                            ETags::VDiskSpace,
+                                            {ETags::PDiskSpace});
+                    break;
+            }
+        }
+
+        storageVDiskStatus.set_overall(context.GetOverallStatus());
+    }
 };
 
 class TBuilderStoragePDisk : public TBuilderBase<TBuilderStoragePDisk> {
+    std::unordered_map<TString, const NKikimrSysView::TPDiskEntry*> PDisksMap;
+
+    TBuilderStoragePDisk(std::unordered_map<TString, const NKikimrSysView::TPDiskEntry*>& pdiskMap)
+        : PDisksMap(pdiskMap)
+    {}
 public:
-    void BuildImpl(const TString& pDiskId, Ydb::Monitoring::StoragePDiskStatus& storagePDiskStatus, TSelfCheckResult context);
+    void FillImpl(const TString& pDiskId, Ydb::Monitoring::StoragePDiskStatus& storagePDiskStatus, TSelfCheckResult context) {
+        context.Location.clear_database(); // PDisks are shared between databases
+        context.Location.mutable_storage()->mutable_pool()->clear_name(); // PDisks are shared between pools
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->clear_id(); // PDisks are shared between groups
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->clear_id(); // PDisks are shared between vdisks
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->Clear();
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->add_pdisk();
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->begin()->set_id(pDiskId);
+        storagePDiskStatus.set_id(pDiskId);
+
+        auto itPDisk = PDisksMap.find(pDiskId);
+        if (itPDisk == PDisksMap.end()) { // this report, in theory, can't happen because there was pdisk mention in bsc vslot info. this pdisk info have to exists in bsc too
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, TStringBuilder() << "System tablet BSC didn't provide expected pdisk information", ETags::PDiskState);
+            storagePDiskStatus.set_overall(context.GetOverallStatus());
+            return;
+        }
+
+        const auto& pDisk = itPDisk->second->GetInfo();
+
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->begin()->set_path(pDisk.GetPath());
+        const auto& statusString = pDisk.GetStatusV2();
+        const auto *descriptor = NKikimrBlobStorage::EDriveStatus_descriptor();
+        auto status = descriptor->FindValueByName(statusString);
+        if (!status) {
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
+                                    TStringBuilder() << "Unknown PDisk state: " << statusString,
+                                    ETags::PDiskState);
+        }
+        switch (status->number()) {
+            case NKikimrBlobStorage::ACTIVE:
+            case NKikimrBlobStorage::INACTIVE: {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+                break;
+            }
+            case NKikimrBlobStorage::FAULTY:
+            case NKikimrBlobStorage::BROKEN:
+            case NKikimrBlobStorage::TO_BE_REMOVED: {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
+                                        TStringBuilder() << "PDisk state is " << statusString,
+                                        ETags::PDiskState);
+                break;
+            }
+        }
+
+        if (pDisk.GetAvailableSize() != 0 && pDisk.GetTotalSize() != 0) { // do not replace it with Has()
+            double avail = (double)pDisk.GetAvailableSize() / pDisk.GetTotalSize();
+            if (avail < 0.06) {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Available size is less than 6%", ETags::PDiskSpace);
+            } else if (avail < 0.09) {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Available size is less than 9%", ETags::PDiskSpace);
+            } else if (avail < 0.12) {
+                context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Available size is less than 12%", ETags::PDiskSpace);
+            }
+        }
+
+        // do not propagate RED status to vdisk - so that vdisk is not considered down when computing group status
+        context.OverallStatus = MinStatus(context.OverallStatus, Ydb::Monitoring::StatusFlag::ORANGE);
+        storagePDiskStatus.set_overall(context.GetOverallStatus());
+    }
 };
 
 class TBuilderStoragePDiskWithWhiteboard : public TBuilderBase<TBuilderStoragePDiskWithWhiteboard> {
+    THashSet<TNodeId> UnavailableStorageNodes;
+
+    TBuilderStoragePDiskWithWhiteboard(THashSet<TNodeId>& unavailableStorageNodes)
+        : UnavailableStorageNodes(unavailableStorageNodes)
+    {}
 public:
-    void BuildImpl(const TString& pDiskId, const NKikimrWhiteboard::TPDiskStateInfo& pDiskInfo, Ydb::Monitoring::StoragePDiskStatus& storagePDiskStatus, TSelfCheckResult context);
+    void FillImpl(const TString& pDiskId, const NKikimrWhiteboard::TPDiskStateInfo& pDiskInfo, Ydb::Monitoring::StoragePDiskStatus& storagePDiskStatus, TSelfCheckResult context) {
+        context.Location.clear_database(); // PDisks are shared between databases
+        if (context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->empty()) {
+            context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->add_pdisk();
+        }
+        context.Location.mutable_storage()->mutable_pool()->clear_name(); // PDisks are shared between pools
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->clear_id(); // PDisks are shared between groups
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->clear_id(); // PDisks are shared between vdisks
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->begin()->set_id(pDiskId);
+        context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk()->begin()->set_path(pDiskInfo.path());
+        storagePDiskStatus.set_id(pDiskId);
+
+        if (pDiskInfo.HasState()) {
+            switch (pDiskInfo.GetState()) {
+                case NKikimrBlobStorage::TPDiskState::Normal:
+                case NKikimrBlobStorage::TPDiskState::Stopped:
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::GREEN);
+                    break;
+                case NKikimrBlobStorage::TPDiskState::Initial:
+                case NKikimrBlobStorage::TPDiskState::InitialFormatRead:
+                case NKikimrBlobStorage::TPDiskState::InitialSysLogRead:
+                case NKikimrBlobStorage::TPDiskState::InitialCommonLogRead:
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW,
+                                            TStringBuilder() << "PDisk state is " << NKikimrBlobStorage::TPDiskState::E_Name(pDiskInfo.GetState()),
+                                            ETags::PDiskState);
+                    break;
+                case NKikimrBlobStorage::TPDiskState::InitialFormatReadError:
+                case NKikimrBlobStorage::TPDiskState::InitialSysLogReadError:
+                case NKikimrBlobStorage::TPDiskState::InitialSysLogParseError:
+                case NKikimrBlobStorage::TPDiskState::InitialCommonLogReadError:
+                case NKikimrBlobStorage::TPDiskState::InitialCommonLogParseError:
+                case NKikimrBlobStorage::TPDiskState::CommonLoggerInitError:
+                case NKikimrBlobStorage::TPDiskState::OpenFileError:
+                case NKikimrBlobStorage::TPDiskState::ChunkQuotaError:
+                case NKikimrBlobStorage::TPDiskState::DeviceIoError:
+                case NKikimrBlobStorage::TPDiskState::Missing:
+                case NKikimrBlobStorage::TPDiskState::Timeout:
+                case NKikimrBlobStorage::TPDiskState::NodeDisconnected:
+                case NKikimrBlobStorage::TPDiskState::Unknown:
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
+                                            TStringBuilder() << "PDisk state is " << NKikimrBlobStorage::TPDiskState::E_Name(pDiskInfo.GetState()),
+                                            ETags::PDiskState);
+                    break;
+                case NKikimrBlobStorage::TPDiskState::Reserved15:
+                case NKikimrBlobStorage::TPDiskState::Reserved16:
+                case NKikimrBlobStorage::TPDiskState::Reserved17:
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Unknown PDisk state");
+                    break;
+            }
+
+            //if (pDiskInfo.HasAvailableSize() && pDiskInfo.GetTotalSize() != 0) {
+            if (pDiskInfo.GetAvailableSize() != 0 && pDiskInfo.GetTotalSize() != 0) { // hotfix until KIKIMR-12659
+                double avail = (double)pDiskInfo.GetAvailableSize() / pDiskInfo.GetTotalSize();
+                if (avail < 0.06) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Available size is less than 6%", ETags::PDiskSpace);
+                } else if (avail < 0.09) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Available size is less than 9%", ETags::PDiskSpace);
+                } else if (avail < 0.12) {
+                    context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Available size is less than 12%", ETags::PDiskSpace);
+                }
+            }
+        } else {
+            if (UnavailableStorageNodes.count(pDiskInfo.nodeid()) != 0) {
+                TSelfCheckResult nodeContext(&context, "STORAGE_NODE");
+                nodeContext.Location.mutable_storage()->clear_pool();
+                nodeContext.Location.mutable_storage()->mutable_node()->set_id(pDiskInfo.nodeid());
+                const TEvInterconnect::TNodeInfo* nodeInfo = nullptr;
+                auto itNodeInfo = MergedNodeInfo.find(pDiskInfo.nodeid());
+                if (itNodeInfo != MergedNodeInfo.end()) {
+                    nodeInfo = itNodeInfo->second;
+                }
+                if (nodeInfo) {
+                    nodeContext.Location.mutable_storage()->mutable_node()->set_host(nodeInfo->Host);
+                    nodeContext.Location.mutable_storage()->mutable_node()->set_port(nodeInfo->Port);
+                }
+                nodeContext.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
+                                            TStringBuilder() << "Storage node is not available",
+                                            ETags::NodeState);
+            }
+            context.ReportStatus(Ydb::Monitoring::StatusFlag::RED,
+                                    TStringBuilder() << "PDisk is not available",
+                                    ETags::PDiskState,
+                                    {ETags::NodeState});
+        }
+
+        storagePDiskStatus.set_overall(context.GetOverallStatus());
+    }
 };
 
 } // NKikimr::NHealthCheck
